Made IniParser static, improved comment writing, and allowed periods in sections.
Changed IniParser so that it is static; changed IniFile to reflect this.

Made writing comments for IniSection and IniSetting better.

Added comments to IniPatternMatcher, and changed the pattern for sections
to allow periods.

Changed the test program to simply open test.ini then save to test2.ini,
so that I can see how the library might change how the saved INIs appear.