Various improvements and fixes.
Hopefully I've solved the stray-comment issue finally, but I doubt it.

There was an issue with comments in the middle of sections. They would be
added to the end of the current INI, rather than in the middle of the
section, so they would appear at the end of the section instead of in the
middle. I fixed this by making sections hold lists of IniParts, just like
IniFiles, and changed the enumerators and indexers like I had to earlier
for IniFile.

I finally remembered how JoinComments worked, and better documented it.
I experimented with using a StringBuilder, but that was actually slower.

IniFile remembers its encoding now so it can use it to write out the file.

Settings have the write-out format "{0}={1}" instead of "{0} = {1}", because
it's more common.