LINQified more code.
IniPatterns.MatchWithPattern is now mostly a LINQ query, which makes it
easier to read and less indented, but slower. I tried using the PLINQ
library, but it made the code MUCH slower. It seems that INI parsing is
too simple a task for it to be worthwhile to setup the threads to
parallelize it, at least for the size of INI files we have.